ai_agents_first/my-chainlit-project/pyproject.toml
[project]
name = "my-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=0.1.0",
    "openai-agents>=0.2.10",
    "python-dotenv>=1.1.1",
]
